Stores an array of @sections instead of a single string in @section
Before `Setting.a.b.c` had `@section == "'b' section in 'a' section in
'[PATH]/settings.yml'"`, now it has `@sections == ['b', 'a',
'Settting']`. The purpose of this commit, is to have the section's
"stack" available and not just a string.

The only use of `@section` was to show a meaningful MissingSetting
message, this is not change in this commit and the message is so
similar, all tests are passing.